+++ src/Mod/Draft/draftguitools/gui_scale.py
@@ -40,12 +40,15 @@ import FreeCAD as App
 import FreeCADGui as Gui
 import Draft_rc
 import DraftVecUtils
+
 import draftutils.utils as utils
+import draftutils.groups as groups
 import draftutils.todo as todo
 import draftguitools.gui_base_original as gui_base_original
 import draftguitools.gui_tool_utils as gui_tool_utils
 import draftguitools.gui_trackers as trackers
 import drafttaskpanels.task_scale as task_scale
+
 from draftutils.messages import _msg, _err
 from draftutils.translate import translate
 
@@ -93,7 +96,8 @@ class Scale(gui_base_original.Modifier):
             self.view.removeEventCallback("SoEvent", self.call)
 
         self.selected_objects = Gui.Selection.getSelection()
-        self.selected_objects = utils.getGroupContents(self.selected_objects)
+        self.selected_objects = \
+            groups.get_group_contents(self.selected_objects)
         self.selected_subelements = Gui.Selection.getSelectionEx()
         self.refs = []
         self.ui.pointUi(self.name)
