+++ src/Mod/Draft/draftutils/utils.py
@@ -591,52 +591,6 @@ def get_clone_base(obj, strict=False):
 getCloneBase = get_clone_base
 
 
-def get_group_names():
-    """Return a list of names of existing groups in the document.
-
-    Returns
-    -------
-    list of str
-        A list of names of objects that are "groups".
-        These are objects derived from `'App::DocumentObjectGroup'`
-        or which are of types `'Floor'`, `'Building'`, or `'Site'`
-        (from the Arch Workbench).
-
-        Otherwise, return an empty list.
-    """
-    glist = []
-    doc = App.ActiveDocument
-    for obj in doc.Objects:
-        if (obj.isDerivedFrom("App::DocumentObjectGroup")
-                or getType(obj) in ("Floor", "Building", "Site")):
-            glist.append(obj.Name)
-    return glist
-
-
-getGroupNames = get_group_names
-
-
-def ungroup(obj):
-    """Remove the object from any group to which it belongs.
-
-    A "group" is any object returned by `get_group_names`.
-
-    Parameters
-    ----------
-    obj : App::DocumentObject
-        Any type of scripted object.
-    """
-    for name in getGroupNames():
-        group = App.ActiveDocument.getObject(name)
-        if obj in group.Group:
-            # The list of objects cannot be modified directly,
-            # so a new list is created, this new list is modified,
-            # and then it is assigned over the older list.
-            objects = group.Group
-            objects.remove(obj)
-            group.Group = objects
-
-
 def shapify(obj):
     """Transform a parametric object into a static, non-parametric shape.
 
@@ -691,139 +645,6 @@ def shapify(obj):
     newobj.Shape = shape
 
     return newobj
-
-
-def get_windows(obj):
-    """Return the windows and rebars inside a host.
-
-    Parameters
-    ----------
-    obj : App::DocumentObject
-        A scripted object of type `'Wall'` or `'Structure'`
-        (Arch Workbench).
-        This will be searched for objects of type `'Window'` and `'Rebar'`,
-        and clones of them, and the found elements will be added
-        to the output list.
-
-        The function will search recursively all elements under `obj.OutList`,
-        in case the windows and rebars are nested under other walls
-        and structures.
-
-    Returns
-    -------
-    list
-        A list of all found windows and rebars in `obj`.
-        If `obj` is itself a `'Window'` or a `'Rebar'`, or a clone of them,
-        it will return the same `obj` element.
-    """
-    out = []
-    if getType(obj) in ("Wall", "Structure"):
-        for o in obj.OutList:
-            out.extend(get_windows(o))
-        for i in obj.InList:
-            if getType(i) in ("Window") or isClone(obj, "Window"):
-                if hasattr(i, "Hosts"):
-                    if obj in i.Hosts:
-                        out.append(i)
-            elif getType(i) in ("Rebar") or isClone(obj, "Rebar"):
-                if hasattr(i, "Host"):
-                    if obj == i.Host:
-                        out.append(i)
-    elif (getType(obj) in ("Window", "Rebar")
-          or isClone(obj, ["Window", "Rebar"])):
-        out.append(obj)
-    return out
-
-
-getWindows = get_windows
-
-
-def get_group_contents(objectslist,
-                       walls=False, addgroups=False,
-                       spaces=False, noarchchild=False):
-    """Return a list of objects from expanding the input groups.
-
-    The function accepts any type of object, although it is most useful
-    with "groups", as it is meant to unpack the objects inside these groups.
-
-    Parameters
-    ----------
-    objectslist : list
-        If any object in the list is a group, its contents (`obj.Group`)
-        are extracted and added to the output list.
-
-        The "groups" are objects derived from `'App::DocumentObjectGroup'`,
-        but they can also be `'App::Part'`, or `'Building'`, `'BuildingPart'`,
-        `'Space'`, and `'Site'` from the Arch Workbench.
-
-        Single items that aren't groups are added to the output list
-        as is.
-
-    walls : bool, optional
-        It defaults to `False`.
-        If it is `True`, Wall and Structure objects (Arch Workbench)
-        are treated as groups; they are scanned for Window, Door,
-        and Rebar objects, and these are added to the output list.
-
-    addgroups : bool, optional
-        It defaults to `False`.
-        If it is `True`, the group itself is kept as part of the output list.
-
-    spaces : bool, optional
-        It defaults to `False`.
-        If it is `True`, Arch Spaces are treated as groups,
-        and are added to the output list.
-
-    noarchchild : bool, optional
-        It defaults to `False`.
-        If it is `True`, the objects inside Building and BuildingParts
-        (Arch Workbench) aren't added to the output list.
-
-    Returns
-    -------
-    list
-        The list of objects from each group present in `objectslist`,
-        plus any other individual object given in `objectslist`.
-    """
-    newlist = []
-    if not isinstance(objectslist, list):
-        objectslist = [objectslist]
-    for obj in objectslist:
-        if obj:
-            if (obj.isDerivedFrom("App::DocumentObjectGroup")
-                    or (getType(obj) in ("Building", "BuildingPart",
-                                         "Space", "Site")
-                        and hasattr(obj, "Group"))):
-                if getType(obj) == "Site":
-                    if obj.Shape:
-                        newlist.append(obj)
-                if obj.isDerivedFrom("Drawing::FeaturePage"):
-                    # skip if the group is a page
-                    newlist.append(obj)
-                else:
-                    if addgroups or (spaces and getType(obj) == "Space"):
-                        newlist.append(obj)
-                    if (noarchchild
-                            and getType(obj) in ("Building", "BuildingPart")):
-                        pass
-                    else:
-                        newlist.extend(getGroupContents(obj.Group,
-                                                        walls, addgroups))
-            else:
-                # print("adding ", obj.Name)
-                newlist.append(obj)
-                if walls:
-                    newlist.extend(getWindows(obj))
-
-    # Clean possible duplicates
-    cleanlist = []
-    for obj in newlist:
-        if obj not in cleanlist:
-            cleanlist.append(obj)
-    return cleanlist
-
-
-getGroupContents = get_group_contents
 
 
 def print_shape(shape):
@@ -1054,64 +875,6 @@ def get_DXF(obj,direction=None):
 
 
 getDXF = get_DXF
-
-
-def get_movable_children(objectslist, recursive=True):
-    """Return a list of objects with child objects that move with a host.
-
-    Builds a list of objects with all child objects (`obj.OutList`)
-    that have their `MoveWithHost` attribute set to `True`.
-    This function is mostly useful for Arch Workbench objects.
-
-    Parameters
-    ----------
-    objectslist : list of App::DocumentObject
-        A single scripted object or list of objects.
-
-    recursive : bool, optional
-        It defaults to `True`, in which case the function
-        is called recursively to also extract the children of children
-        objects.
-        Otherwise, only direct children of the input objects
-        are added to the output list.
-
-    Returns
-    -------
-    list
-        List of children objects that have their `MoveWithHost` attribute
-        set to `True`.
-    """
-    added = []
-    if not isinstance(objectslist, list):
-        objectslist = [objectslist]
-    for obj in objectslist:
-        # Skips some objects that should never move their children
-        if getType(obj) not in ("Clone", "SectionPlane",
-                                "Facebinder", "BuildingPart"):
-            children = obj.OutList
-            if hasattr(obj, "Proxy"):
-                if obj.Proxy:
-                    if (hasattr(obj.Proxy, "getSiblings")
-                            and getType(obj) not in ("Window")):
-                        # children.extend(obj.Proxy.getSiblings(obj))
-                        pass
-            for child in children:
-                if hasattr(child, "MoveWithHost"):
-                    if child.MoveWithHost:
-                        if hasattr(obj, "CloneOf"):
-                            if obj.CloneOf:
-                                if obj.CloneOf.Name != child.Name:
-                                    added.append(child)
-                            else:
-                                added.append(child)
-                        else:
-                            added.append(child)
-            if recursive:
-                added.extend(getMovableChildren(children))
-    return added
-
-
-getMovableChildren = get_movable_children
 
 
 def filter_objects_for_modifiers(objects, isCopied=False):
